Mongsil-4 [architecture] model layer 리팩토링하기
- LocalDataSource 추가
- ViewModel 분리
