--- app/src/main/java/com/cashproject/mongsil/ui/calendar/day/SayingCase.kt
@@ -1,5 +1,0 @@
-package com.cashproject.mongsil.ui.calendar.day
-
-enum class SayingCase{
-    CALENDAR, LIST
-}
