+++ app/src/main/java/com/cashproject/mongsil/ui/pages/calendar/day/SayingCase.kt
@@ -1,0 +1,5 @@
+package com.cashproject.mongsil.ui.pages.calendar.day
+
+enum class SayingCase{
+    CALENDAR, LIST
+}
