+++ app/src/main/java/com/cashproject/mongsil/ui/pages/locker/LockerAdapter.kt
@@ -1,29 +1,24 @@
-package com.cashproject.mongsil.ui.main
+package com.cashproject.mongsil.ui.pages.locker
 
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
-import com.cashproject.mongsil.databinding.ItemCommentBinding
-import com.cashproject.mongsil.model.data.Comment
+import com.cashproject.mongsil.databinding.ItemLockerBinding
+import com.cashproject.mongsil.model.data.LikeSaying
 
-class CommentAdapter : RecyclerView.Adapter<CommentAdapter.ViewHolder>() {
+class LockerAdapter : RecyclerView.Adapter<LockerAdapter.ViewHolder>() {
 
-    private var items: ArrayList<Comment> = ArrayList()
+    private var items: ArrayList<LikeSaying> = ArrayList()
 
-    private var listener: ((item: Comment) -> Unit)? = null
+    private var listener: ((item: LikeSaying) -> Unit)? = null
 
-    fun setOnItemClickListener(listener: (item: Comment) -> Unit) {
+    fun setOnItemClickListener(listener: (item: LikeSaying) -> Unit) {
         this.listener = listener
     }
 
-    fun setItems(items: ArrayList<Comment>){
-        this.items = items
-        notifyDataSetChanged()
-    }
-
-    fun update(newItemList: List<Comment>) {
+    fun update(newItemList: List<LikeSaying>) {
         val diffResult = DiffUtil.calculateDiff(ContentDiffUtil(items, newItemList), false)
         diffResult.dispatchUpdatesTo(this)
         items.clear()
@@ -32,27 +27,30 @@ class CommentAdapter : RecyclerView.Adapter<CommentAdapter.ViewHolder>() {
 
     override fun getItemCount() = items.size
 
-    override fun onBindViewHolder(holder: CommentAdapter.ViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: LockerAdapter.ViewHolder, position: Int) {
         holder.bindView(items[position])
         Log.d("Locker", items[position].toString())
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder
-     = ViewHolder(ItemCommentBinding.inflate(LayoutInflater.from(parent.context), parent, false))
+     = ViewHolder(ItemLockerBinding.inflate(LayoutInflater.from(parent.context), parent, false))
 
 
-    inner class ViewHolder(private val binding: ItemCommentBinding) : RecyclerView.ViewHolder(binding.root){
+    inner class ViewHolder(private val binding: ItemLockerBinding) : RecyclerView.ViewHolder(binding.root){
 
-        fun bindView(item: Comment){
-            binding.comment = item
+        fun bindView(item: LikeSaying){
+            binding.saying = item
+            binding.ivLockerImage.setOnClickListener{
+                listener?.invoke(item) //익명함수 호출
+            }
         }
     }
 
-    class ContentDiffUtil(private val oldList: List<Comment>, private val currentList: List<Comment>) : DiffUtil.Callback() {
+    class ContentDiffUtil(private val oldList: List<LikeSaying>, private val currentList: List<LikeSaying>) : DiffUtil.Callback() {
 
         //1. 아이템의 고유 id 값이 같은지
         override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-            return oldList[oldItemPosition].id == currentList[newItemPosition].id //title 대신 id 사용가능
+            return oldList[oldItemPosition].docId == currentList[newItemPosition].docId //title 대신 id 사용가능
         }
 
         //2. id 가 같다면, 내용물도 같은지 확인 equals()
